Fix bug in myeeprom.h - frskyVoltThreshold wrong size

+++ trunk/src/myeeprom.h
@@ -222,7 +222,7 @@ typedef struct t_ModelData {
     int8_t    curves5[MAX_CURVE5][5];
     int8_t    curves9[MAX_CURVE9][9];
     CSwData   customSw[NUM_CSW];
-    uint16_t  frSkyVoltThreshold ;
+    uint8_t   frSkyVoltThreshold ;
     uint8_t   res3[2];
     SafetySwData  safetySw[NUM_CHNOUT];
     FrSkyData frsky;
